BackEnd/healthcare-backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
argon2-cffi==23.1.0

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Utilities
python-dotenv==1.0.0
cryptography==42.0.0

# Time-series (TimescaleDB is a PostgreSQL extension)
# No additional Python package needed

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
black==23.12.1
flake8==7.0.0
email-validator>=2.0.0
